requirements.txt
openai-whisper==20231117
moviepy==1.0.3
deep-translator==1.11.4